Update site finding to use "find desired location"
This is a good place to define what we require in
the guide and what we do not. If we are requiring
soil, then we can include a section on initial
soil farming (we don't need a branching guide that
must account for both). Same with clay and 
pottery.

Removed the long bit about DFHack and Legends
mode, I think it's out of scope (maybe add as
an appendix?)